// shopcart/src/app/types/product.ts
export interface ProductReview {
  rating: number;
  comment: string;
  date: string;
  reviewerName: string;
  reviewerEmail: string;
}

export interface Product {
  _id: string;

  title: string;
  description: string;
  category: string;

  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;

   isActive: boolean;

  tags?: string[];
  brand?: string;
  sku?: string;
  weight?: number;

  dimensions?: {
    width: number;
    height: number;
    depth: number;
  };

  warrantyInformation?: string;
  shippingInformation?: string;
  availabilityStatus?: string;

  reviews?: ProductReview[];

  returnPolicy?: string;
  minimumOrderQuantity?: number;

  meta?: {
    createdAt: string;
    updatedAt: string;
    barcode: string;
    qrCode: string;
  };

  images: string[];
  thumbnail: string;
}

export interface ProductsResponse {
  products: Product[];
  total: number;
  skip: number;
  limit: number;
}